make gcc happy
(Logical change 1.2305)

+++ plugin/item/static_stat.c
@@ -330,7 +330,7 @@ static int unix_sd_present( struct inode *inode /* object being processed */,
 		inode -> i_mtime = d32tocpu( &sd -> mtime );
 		inode -> i_ctime = d32tocpu( &sd -> ctime );
 		inode -> i_rdev  = val_to_kdev( d32tocpu( &sd -> rdev ) );
-		inode_set_bytes( inode, d64tocpu( &sd -> bytes ) );
+		inode_set_bytes( inode, ( loff_t ) d64tocpu( &sd -> bytes ) );
 		move_on( len, area, sizeof *sd );
 		return 0;
 	} else
